test(eval-longmemeval): move warm-create speed gate to slow tier with p75/n=20
Speed-gate test flaked intermittently in the parallel suite (PGLite I/O
contention pushed p50 over 500ms occasionally). Two independent expert
reviewers (performance-engineer, test-automator) recommended A+E:

A — Move to .slow.test.ts. The slow tier runs sequentially via
scripts/run-slow-tests.sh, outside parallel shard contention. Uses
the existing repo tier convention (parallel / serial / slow).

E — Tighten statistical gate. Bump trials 10 → 20 and assert p75
< 500 instead of p50. Reasoning: in the slow tier we have headroom
(60s/test, sequential), so we can afford a more robust gate. p75 of
20 isolated trials catches a genuine 2x regression reliably; p50 of
10 contended trials was statistically thin and flaky.

The 500ms threshold is unchanged. p99 warn-only block preserved.
The correctness tests in eval-longmemeval.test.ts stay in the
parallel suite as before.
